# Conflicts:
#	angular-frontend/src/app/components/navbar/navbar.component.html

+++ src/main/java/hu/progmasters/fundraiser/dto/fund/FundFormCommand.java
@@ -1,7 +1,5 @@
 package hu.progmasters.fundraiser.dto.fund;
 
-
-
 import java.time.LocalDate;
 
 public class FundFormCommand {
@@ -16,10 +14,9 @@ public class FundFormCommand {
 
     private String category;
 
-    private Integer targetAmount;
+    private Double targetAmount;
 
     private LocalDate endDate;
-
 
     public static FundFormCommand getDummyInstance(String name) {
         FundFormCommand instance = new FundFormCommand();
@@ -28,7 +25,7 @@ public class FundFormCommand {
         instance.setLongDescription("Long Descripton tow");
         instance.setImageUrl("image_two.jpg");
         instance.setCategory("MEDICAL");
-        instance.setTargetAmount(1001);
+        instance.setTargetAmount(1001.0);
         instance.setEndDate(LocalDate.now());
         return instance;
     }
@@ -53,7 +50,7 @@ public class FundFormCommand {
         return category;
     }
 
-    public Integer getTargetAmount() {
+    public Double getTargetAmount() {
         return targetAmount;
     }
 
@@ -81,11 +78,12 @@ public class FundFormCommand {
         this.category = category;
     }
 
-    public void setTargetAmount(Integer targetAmount) {
+    public void setTargetAmount(Double targetAmount) {
         this.targetAmount = targetAmount;
     }
 
     public void setEndDate(LocalDate endDate) {
         this.endDate = endDate;
     }
+
 }
