+++ src/main/java/hu/progmasters/fundraiser/dto/transfer/create/TransferCreationCommand.java
@@ -11,21 +11,17 @@
 
 package hu.progmasters.fundraiser.dto.transfer.create;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
-
 public class TransferCreationCommand {
 
     private Long targetFundId;
 
-    private Integer amount;
+    private Double amount;
 
     public Long getTargetFundId() {
         return targetFundId;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
@@ -33,7 +29,7 @@ public class TransferCreationCommand {
         this.targetFundId = targetFundId;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
